Support getting and setting variant name. Closes #743
